Feat: Cambio en la estructura de la página de Dashboard
